Authorization config

+++ App/controllers/ListingController.php
@@ -3,6 +3,8 @@
 namespace App\Controllers;
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 
 class ListingController {
 
@@ -16,9 +18,7 @@ class ListingController {
 
     public function index() {
 
-        /* inspectAndDie(Validation::email('test$test.com')); */
-
-        $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
         loadView('listings/index', [
         'listings' => $listings
@@ -84,7 +84,7 @@ class ListingController {
          */
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map('sanitize', $newListingData);
 
@@ -150,9 +150,17 @@ class ListingController {
 
         $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
 
+        //Check if listing exists
         if(!$listing) {
             ErrorController::notFound('Listing not found');
             return;
+        }
+
+        //Authorize user
+
+        if(!Authorization::isOwner($listing->user_id)) {
+            $_SESSION['error_message'] = 'You are not authorized to delete this listing';
+            return redirect('/listings/' . $listing->$id);
         }
 
         $this->db->query('DELETE FROM listings where id = :id', $params);
